Clang-Tidy VI
* Update GNU Make.  One can use `USE_CLANG_TIDY=TRUE` to enable clang-tidy
  check.  For example, to enable clang-tidy check, specify the clang-tidy
  command (if it's the default clang-tidy), and treat warnings as errors,
  one can run

      make USE_CLANG_TIDY=TRUE CLANG_TIDY=clang-tidy-12 CLANG_TIDY_WARN_ERROR=TRUE

  By default amrex/.clang-tidy is used as the config file.  One can change
  it with CLANG_TIDY_CONFIG_FILE=my_clang-tidy-conf.  If you want to let
  clang-tidy fix errors for you, you can run

      make USE_CLANG_TIDY=TRUE CLANG_TIDY="clang-tidy --fix-errors"

  To avoid race conditions, the make job should not use `-j` (unless it's
  `-j1`) when fixing errors.

* Add clang-tidy to GNU Make CIs.

* Fix warnings in Fortran interface, particle and plotfile tools.

+++ Tools/Plotfile/fcompare.cpp
@@ -137,7 +137,7 @@ int main_main()
             amrex::Print() << " WARNING: variable " << names_a[n_a] << " not found in plotfile 2\n";
             all_variables_found = false;
         } else {
-            ivar_b[n_a] = std::distance(std::begin(names_b), r);
+            ivar_b[n_a] = static_cast<int>(std::distance(std::begin(names_b), r));
         }
 
         if (names_a[n_a] == diffvar) {
@@ -369,7 +369,7 @@ int main_main()
         if (abort_if_not_all_found) return EXIT_FAILURE;
     }
 
-    if (any_nans) {
+    if (any_nans) { // NOLINT(bugprone-branch-clone)
         return EXIT_FAILURE;
     } else if (global_error == 0.0) {
         amrex::Print() << " PLOTFILE AGREE" << std::endl;
